+++ modules/core-config/modules/node-config/variables.tf
@@ -1,0 +1,19 @@
+variable "namespace" {
+  description = "Namespace to install the Kubernetes resources into."
+  type        = string
+  nullable    = false
+}
+
+variable "labels" {
+  description = "Labels to be applied to all Kubernetes resources."
+  type        = map(string)
+  nullable    = false
+}
+
+variable "timeouts" {
+  description = "Timeout configuration."
+  type = object({
+    helm_modify = number
+  })
+  nullable = false
+}