+++ modules/node-groups/modules/bootstrap-node-group-hack/variables.tf
@@ -16,6 +16,12 @@ variable "cluster_name" {
   nullable    = false
 }
 
+variable "fips" {
+  description = "If true, the cluster will be created with FIPS 140-2 mode enabled; this can't be changed once the cluster has been created."
+  type        = bool
+  nullable    = false
+}
+
 variable "subnet_id" {
   description = "ID of the subnet to use for the bootstrap node group."
   type        = string
@@ -31,11 +37,5 @@ variable "bootstrap_name" {
 variable "bootstrap_vm_size" {
   description = "VM size to use for the bootstrap node group."
   type        = string
-  nullable    = false
-}
-
-variable "fips" {
-  description = "If true, the cluster will be created with FIPS 140-2 mode enabled; this can't be changed once the cluster has been created."
-  type        = bool
   nullable    = false
 }